+++ App/Next/next_app/src/app/firstLogin/page.tsx
@@ -1,18 +1,19 @@
 "use client";
 import Image from 'next/image';
 import CustomBtn from "@/components/CustomBtn";
+import Manage2FAFirstLogin from "@/components/auth/2faFirstLoginAuth";
 
 export default function FirstLogin() {
     return (
         <div className="flex flex-col items-center justify-center">
             <div className="m-4 pt-4">
                 <p className="font-bold text-center">Choose your username</p>
-                <input type="text" className="m-2 bg-base border-red  border-0  w-64 h-8 focus:outline-none"/>
+                <input type="text" className="m-2 bg-base border-red  border-0  w-64 h-8 focus:outline-none" />
             </div>
-                
+
             <div className="m-4 flex-auto">
                 <p className="font-bold mb-2">Choose your avatar</p>
-                <Image 
+                <Image
                     src="https://img.freepik.com/free-icon/user_318-563642.jpg"
                     alt="default avatar"
                     width={128}
@@ -20,18 +21,12 @@ export default function FirstLogin() {
                     className=" drop-shadow-xl"
                 />
             </div>
-            
-            <div className="m-4 flex flex-auto items-center">
-                <label className=" h-4">
-                    <input type="checkbox" 
-                        name="2fa" id="2fa" 
-                        className="w-4 h-4 mr-2  accent-mauve before:accent-lavender unchecked-checkbox"/>
-                    Activate Two-Factor
-                </label>
+            <div className="flex flex-col flex-auto items-center justify-center">
+                <Manage2FAFirstLogin></Manage2FAFirstLogin>
             </div>
             <CustomBtn onClick={() => console.log("Validate!")}>
                 Validate
             </CustomBtn>
         </div>
     )
-  }
+}
